requirements.txt
python-dotenv
streamlit
langchain-community
langchain-text-splitters
langchain-groq
langchain-huggingface
unstructured
unstructured[pdf]
faiss-cpu
transformers
sentence-transformers
nltk